Do adaptive hashing
... that is, change the hash function during the lifetime of a hash
table to make it faster.

- For EQ and EQL hash tables up to 16 buckets, use a single vector and
linear search. This is 15-40% faster. This is adaptive hashing only
inasmuch as it can be viewed as a constant hash.

- Above that range, for EQ hash tables only, a very simple hash
function is used until the number of collisions gets too high, at
which point we switch to the previously default hash function. This is
a few percents faster and lays the groundwork for fancier scheme and
also for preventing denial-of-service attacks by an adversarial key
generator.

- EQUAL and EQUALP hash table now increase how deep and wide they hash
conses if there are two many collisions. For example, this makes
tests/arith-combinations.lisp 50% faster because it uses list keys
whose first +MAX-HASH-DEPTHOID+ (4) elements are constant.