Rehash only if the INDEX-VECTOR grows
When a hash table grows, its INDEX-VECTOR may not grow because its
length is basically the maximum capacity rounded up to a power of 2.
In fact, with the default REHASH-SIZE being 1.5, this happens in about
4/9 of the cases.

This change optimizes this case by not reallocating the INDEX-VECTOR
and simply copying over NEXT-VECTOR contents instead of doing a full
REHASH. This is made possible by the stickiness of the address-based
flag (SB-VM:VECTOR-ADDR-HASHING-FLAG).