Adaptive hashing: flat hash tables
For EQ and EQL hash tables up to 32 and 16 buckets, respectively, use
a single vector and linear search without hashing. Above that number
of buckets, switch to the normal hash table implementation.

Flat hash tables do not use the index/next vectors, and do not
interact with GC. They are also 15-40% faster.

This is adaptive hashing only inasmuch as it can be viewed as a
constant hash with a specialized implementation.